Integ tests failed to create a control socket
The RUN_SOCKET test was failing while trying to make the control socket because
it tried to change the permissions of /tmp to 700 (which, of course, normal
users can't).

Tor rejects to start if the control socket is in a globally readable directory,
so making a temporary directory we can set the permissions on.