Safecookie authentication broken when using python 3
The os.urandom() provides bytes so there's no need to convert that, but we
needed to be more careful about using the converted hex in the formatted
string. It too was bytes, causing an extra embedded b''...

>>> "hello %s" % b"damian"
"hello b'damian'"

Tor in turn rejected the nonce as being invalid. I don't get why python3 does
it. With python2 bytes and unicode could both be used in a reasonable fashion
for formatted strings - these extra characters are almost always undesirable.
Oh well...