Still passing when tor's stime is zero
The stime proc field is an integer, rounded down to the number of seconds that
a process has been running. On my piddly little netbook tor takes long enough
to start that this is always non-zero, but on our jenkins systems it's so quick
that the test fails. Thanks to weasel for all of the help!