Dynamically importing test modules
Whenever we add a new test we need to add it to both the imports and UNIT_TESTS
or INTEG_TESTS tuple. This is both a little clunky and, now that we have a crap
ton of tests, it makes run_tests.py really long.

Using __import__() to dynamically import the tests as we run them. This might
have a slight runtime advantage too since we're just importing what we need,
though it's not noticeable. This trick comes from...

http://effbot.org/zone/import-string.htm