Checking for unused tests
Well... shame on me. We've always had a hardcoded list of our tests, but
somehow when I moved it to our settings.cfg I missed a few. I'm not so much
troubled about that mistake, but more that this mistake went undetected.

Adding a check at the start of our tests for this kind of misconfiguration.