Raising FAQ category types
Changing the category types so...

'subcategories' => 'categories'
'subsubcategories' => 'subcategories'

This better emphasizes the questions and gives them a slight horizontal ruler.
I'd like to get a more emphasized divider to better split up the questions but
the only way I know of to do that is a table, which would be gross.