Bumping sleep for test_reattaching_listeners
On our jenkins instance we're seeing this test fail roughly 1/24 of the time
(breaking one target in 1/3 of the test runs). We *should* be doing some kind
of activity guarenteed to trigger an event rather than rely on a BW event being
emitted each second. For now trying to simply bump its sleeps though to see if
that relieve the symptoms.

This adds around eight seconds to our integ runtime so definitely something we
should look into doing right later.