Support for pulling multiple pids with get_pid_by_name()
A leading cause of test failures in our jenkins environment is
test_get_pid_by_name_lsof. This is because the test assumes that there is only
a single tor process, but that's often not the case on the dixie host.

This not only fixes the test, but expands get_pid_by_name() to include a
'multiple' flag to pull all pids for a given process name. This includes unit
test coverage.