Normalize ControlMessage content to be bytes
The socket module's recv_message function converted read content to be a str
(bytes in python 2.x and unicode in python 3.x). In 99.9% of the cases this is
exactly what we want, but in a few edge cases we actually want byte content
under python 3.x.

For instance, in #8755 descriptors with non-unicode contact lines are being
altered by the socket module, causing our validation of their signature to
fail.

Functionally this commit should have no effect besides adding the option to get
byte content from the ControlMessage. The tests seem happy so fingers crossed!