Processing controller descriptors as bytes
When descriptor content was fetched by the Controller it was first converted to
unicode, then back into bytes. As a result descriptors with content that
couldn't be converted (usually on the contact line) got mangled, and this in
turn caused our check of its signature to fail.

Caught thanks to aj00200 on...

https://trac.torproject.org/8755