System module's call() function ignored exit statuses
Well... oops. Our call() method was documented as raising an OSError (or
returning the default value) when the call fails but it completely ignored the
exit status. In practice the function only behaved as documented when Popen
failed - otherwise it returned an empty list. Fixing this behavior and updating
most of our callers since they expected call() to not raise any exceptions.