Adding the mock module as a test dependency
Mocking is hard. We've gotten a lot of mileage by writing our own mocking
helpers, but they're clunky and complicated. It's about time for us to use a
more standard, mature option.

The pypi mock module is not only well documented, but looks to be the direction
the python world is going. As of python 3.3 it's now part of the standard
library.

As the first step of swapping us over I'm adding it as a dependency for
run_tests.py. This warns the user if it's unavailable...

  atagar@morrigan:~/Desktop/stem$ ./run_tests.py --unit
  To run stem's tests you'll need mock...

  https://pypi.python.org/pypi/mock/

  You can get it by running 'sudo pip install mock'.