Using mock for tutorial unit tests
The unit tests for our tutorials are some of the ugliest we have since they
make heavy use of mocks. Hence a good next step for mock.

The mock module makes the controller mocking considerably nicer. That said,
mock's open() mocking is crap. Their mock_open() docs...

http://www.voidspace.org.uk/python/mock/helpers.html#mock.mock_open

Tripped me up for well over an hour because their examples are hardcoded for
the __main__ namespace (so it works when I do it in an interpretor, but not the
tests). Even with that fixed the mock object it returns is buggy as hell (it's
documented as supporting readlines() but that's a lie). They'd get a lot more
mileage if they used an io.BytesIO instead.