Clarifying why test's load_conf() expects an exception
For a long while I've been puzzled why this load_conf() call asserted an
exception. When I ran it in a cusom prompt it ran fine - turned out that it was
relying on us having a custom DataDir. Tor balks (rightfully) when we'd shift
our DataDir out from a running tor instance.

Doing a better job explaining what's up in the comment.