Workaround for screwy RuntimeError related to mock causing imports
The following isn't used by this directly, but we're still importing it due
to a screwy aspect of how mock works. If patched() results in an import that
we haven't done before then we can fail with...

  ======================================================================
  ERROR: test_query_download
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "/usr/local/lib/python2.7/dist-packages/mock.py", line 1201, in patched
      return func(*args, **keywargs)
    ...
    File "/home/atagar/Desktop/stem/stem/descriptor/__init__.py", line 156, in parse_file
      import stem.descriptor.networkstatus
  RuntimeError: cannot unmarshal code objects in restricted execution mode

I've only seen this arise when running this test directly (without the rest
of the unit tests, some of which import networkstatus on their own).