Fixing "cannot unmarshal code" errors
Damnit python, your import scheme is stupidly confusing.

The descriptor's __init__ module has a circular dependency with its contents.
This is because the parse_file() function calls the constituent modules, while
those modules need the Descriptor class from __init__.

So far so good. Only trouble is that python's support for circular dependencies
sucks. To address this I did lazy imports in __init__, so we imported within
the parse_file() function.

On the surface this seemed to work. All the tests certainly passed. The trouble
is that this style of python import is buggy as hell. Turns out that lazy
imports leave the module in question in a unexecutable state so this *only*
works if you've also imported the module another time during the interpretor
execution. Our tests did this, hence passing tests.

I first encuntered "cannot unmarshal code" while writing the remote descritpor
tests (both unit and integ). I was content to hack around this with superfluous
import statements while this only manifested within the tests, but now I'm
seeing it during general usage too...

>>> from stem.descriptor.remote import DescriptorDownloader
>>> list(DescriptorDownloader().get_microdescriptors('jzcx+1fHsi47Tu+vQIcyItgn4lKs6aKnFshQ0lZ2JTg'))
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "stem/descriptor/remote.py", line 311, in __iter__
    for desc in self.run(True):
  File "stem/descriptor/remote.py", line 300, in run
    for desc in self._results:
  File "stem/descriptor/__init__.py", line 154, in parse_file
    import stem.descriptor.server_descriptor
RuntimeError: cannot unmarshal code objects in restricted execution mode

Joy. After much head scratching and forum reading it sounds like there's
something magical about 'from' imports so switching the descriptor modules to
that, and moving the __init__ imports to the end. I'm not entirely clear on the
magic going on here, but its elmiminated the errors.