Having use_directory_mirrors() return the consensus
Originally I was planning an on-disk cache so calling get_consensus() after
use_directory_mirrors() would result in an on-disk read. However, that has been
put on hold for now so having use_directory_mirrors() return the consensus it
fetches in order to get the mirrors.

We unfortunately already need to read the full consensus into memory before
processing it due to how urllib2 works, so this shouldn't cause a memory usage
hit. We might someday want to change this though if we get a constant time
download/parsing scheme.