Defaulting Query's fall_back_to_authority to False
When a user opts to use the Query class direcly they're doing so to use
specific endpoints. We'll probably cause more confusion then help by falling
back to the authorities here. We'll still fall back by default in the
downloader, though.

This also includes a tweak so users can override the downloader's endpoints.