Query's run() method didn't block
Well, that was confusing as hell. Say you wrote the following script...

  query = stem.descriptor.remote.Query(
    '/tor/server/all.z',
    timeout = 60,
  )

  query.run(True)

  if not query.error:
    print "Results downloaded: %i" % len(list(query))
  else:
    print "Got an error: %s" % query.error

You'd expect run() to block, right? You should - that's what it was documented
as doing. Instead it returned an iterator and we don't block until the
list(query) call. This means that we would *never* report an error. This has
been causing stacktraces for my monitors every time it encountered an error.