Adding a 'content' attribute to the Query class
Including a Query attribute so the caller can get our raw descriptor content.
I'm actually doing this to simplify the run() method's usage. It had provided
an iterator so...

  query = Query(my_resource)
  print list(query)  # this would print the ist of descriptors
  print list(query)  # this would be an empty list

The reason the second call is an empty list is because we already iterated over
the query. This is confusing, especially since...

  query = Query(my_resource)
  query.run(True)
  print list(query)  # also an empty list

... due to run() returning a list under the covers. By making run() provide a
fresh iterator each time I also now had the downloaded content handy so making
it public made sense (it'll commonly be wanted by our callers).