Propagating additional descriptor keyword arguments
Adding a 'kwargs' argument to all methods for fetching descriptors. This
argument is passed along to the descriptor contructor. I just ran into a need
for this because the NetworkStatusDocumentV3 class has a 'default_params'
argument that I couldn't access through parse_file(). We'll likely be adding
more constructor args in the future so this will help to make those... well,
useful.