Don't fail tests upon lingering thread
The chief cause of our present jenkins failures are lingering threads...

https://trac.torproject.org/projects/tor/ticket/8896

This is certainly an problem. It indicates either a bug with stem closing
threads, or that we aren't closing a controller in the tests. However, I'm
tired of the false alarms from the jenkins tests, besides which this is a very
minor issue.