Handling bridge-ip-versions as a general extrainfo field
In commit 0d01ad6 I added support for the 'bridge-ip-versions' lines. These are
only relevant to bridge descriptors, and ineed don't presently appear in the
extrainfo consensus. However, that once could be said about the 'bridge-ips'
field too.

It's wrong for bridge attributes to appear on public relays, but there's
nothing in the spec prohibiting it. Hence moving this attribute's handler to
the common ExtraInfoDescriptor parent class.