Adding tor directory authority information
To work our remote descriptor fetching module needed the addresses and dirports
of voting authorities. However, on reflection it's useful if we reflect the
rest of the authority information from tor's config.c as well. For instance,
the v3ident is handy if our users want to query authority key certificates.