Replacing DIRECTORY_AUTHORITIES constant with get_authorities()
On reflection a DIRECTORY_AUTHORITIES constant is not something I want to vend
in our next stem release. That would make it tricky to be flexible about where
we get that information. For instance, we might want to get this from tor
itself (if that ever becomes an option), or possibly a config file rather than
hardcoded dict.