Disable caching of HiddenServiceOptions
Stem's controller has pretty straight forward caching for tor's configuration
options. If you call 'GETCONF foo' then the value of foo is cached until you
call either 'SETCONF foo' or 'RESETCONF foo'.

In general this is all well and good, but HiddenServiceOptions behave in a
different fashion. With those the caller might call 'SETCONF HiddenServiceDir'
and expect 'GETCONF HiddenServiceOptions' to change. As such, opting out of
caching hidden service options.

This was a bug uncovered by wayzard on...

http://stackoverflow.com/questions/18777837/tors-stem-module-python-loading-but-not-unloading-config
https://trac.torproject.org/projects/tor/ticket/9792

Prior to this fix...

>>> from stem.control import Controller
>>> controller = Controller.from_port()
>>> controller.authenticate()
>>> controller.get_conf_map('HiddenServiceOptions')
{'HiddenServiceOptions': []}
>>> controller.set_options([('HiddenServiceDir','/tmp/my_hidden_service'), ('HiddenServicePort', '1236 127.0.0.1:1236')])
>>> controller.get_conf_map('HiddenServiceOptions')
{'HiddenServiceOptions': []}

Now it behaves as expected...

>>> from stem.control import Controller
>>> controller = Controller.from_port()
>>> controller.authenticate()
>>> controller.get_conf_map('HiddenServiceOptions')
{'HiddenServiceOptions': []}
>>> controller.set_options([('HiddenServiceDir','/tmp/my_hidden_service'), ('HiddenServicePort', '1236 127.0.0.1:1236')])
>>> controller.get_conf_map('HiddenServiceOptions')
{'HiddenServicePort': ['1236 127.0.0.1:1236'], 'HiddenServiceDir': ['/tmp/my_hidden_service']}