Halo for windrose icon
Adding a blue halo for the center of the windrose. This doesn't address the
main thing I dislike about the icon, which is the pixelation of the black
lines, but still it makes it look a bit nicer. I've tried several things to
address the black line pixilation (blur, thickening, etc) but this is the best
improvement I've yet found.