Adding python 2.x backport of @lru_cache
Python 3.2 added a memoization annotation to python's functools module. This is
very, very handy, allowing us to avoid the caching boilerplate I do way too
often...

  def get_foo():
    if self._foo is None:
      ... stuff to calculate self._foo...

    return self._foo

With a memoization function this becomes...

  @lru_cache()
  def get_foo():
    ... stuff to calculate self._foo...

This is a MIT licensed backport from...

  http://code.activestate.com/recipes/578078-py26-and-py30-backport-of-python-33s-lru-cache/

Looking forward to when we require python 3.2 so we can use the builtin!