Adding hash functions to exit policy classes
Huh, didn't expect that. To work the @lru_cache requires that all arguments
(including the class in the ase of methods) are hashable. This makes perfectly
sense, after all the cache is a 'argument => cached value' dictionary.

Under python 2.x all seemed to be well, but under python 3.x the @lru_cache
complained that our classes weren't hashable (which... well, they aren't). Why,
then, did it work under python 2.x? Turns out there's a subtle difference where
our object parent provides a __hash__ method for our id, but in python 3.x it's
a little different (object still does, but for reasons I haven't been able to
figure out it doesn't for our classes).

The ExitPolicyRule can certainly be hashable. The ExitPolicy, however, is a lot
trickier due to its lazy loading of rules. For now just opting for the same
behavior as python 2.x and using the address.