Moving static checks to the start
Rearranging our test run so the static checks (pyflakes and pep8) occure at the
start and are listed among the initialization tasks. This makes it clearer that
they're running and the runtime dedicated to them. It also gets rid of a silent
pause at the end of the test run I've always disliked.