Moving pep8 and pyflakes helpers to stem.util.test_tools
Twice now I've copied the part of stem's test/util.py for performing pyflakes
and pep8 checks (for both arm and ExitMap). I don't want to vend this
functionality but it's silly for us to keep copying this functionality.

Generalizing the helper functions and moving them to our util. Like the log
module, this has a note that it is not something we vend and that users should
not rely on it to have a backward compatible API.