Lowering logging about undelivered messages to the info level
Our BaseController's test_msg_repeatedly integ test hammers the holy bejezus
out of our controller socket in the hope of triggering exotic concurrency
issues that might lead to issues like deadlock. This has been an impressively
great test at ferreting out exotic issues and this is no different. Our Jenkins
tests frequently log...

  03/12/2014 15:37:57 [NOTICE] BUG: the msg() function failed to deliver a
    response: Authentication required.

After a couple hours of looking into this I gotta admit that I'm frustrated by
the delicacy of our message handling. It works great in practice, but in theory
issues such as socket errors or edge cases can cause undelivered messages.

In this case rapidly connecting and closing the socket while simultaneously
communicating on it causes the socket's recv() to occasionally raise
exceptions. These exceptions are reported to callers, leaving enqueued messages
that end up being undelivered (logging this notice).

After much head scratching and digging into our socket class I've come to the
realization that this is about as good as it's gonna get. If we use our recv
queue to communicate *both* messages and exceptions then it is to be expected
that exceptions can sometimes result in undelivered messages.

Lowering the severity of the log message appropriately.