Invalid event types caused add_event_listener() error
When presented with invalid or unrecognized event types our
add_event_listener() is documented as raising a ProtocolError. However, it
caused a KeyError instead...

  >>> SETEVENTS BW,DEBUG
  Traceback (most recent call last):
    File "./prompt", line 8, in <module>
      stem.interpretor.main()
    File "/home/atagar/Desktop/stem/stem/interpretor/__init__.py", line 67, in main
      print interpretor.run_command(user_input)
    File "/home/atagar/Desktop/stem/stem/interpretor/commands.py", line 449, in run_command
      self.controller.add_event_listener(self.register_event, *events)
    File "/home/atagar/Desktop/stem/stem/control.py", line 1920, in add_event_listener
      event_version = stem.response.events.EVENT_TYPE_TO_CLASS[event_type]._VERSION_ADDED
  KeyError: 'BW,DEBUG'

Skipping the version check if we don't have a class for the event type. This
both fixes invalid events, and also allows us to proper new event types that
Stem doesn't recognize yet.

  >>> SETEVENTS BW,DEBUG
  SETEVENTS rejected BW,DEBUG