Caching autocompletion matches
Readline's autocompletion is kinda weird. It uses a function of the form...

  matches(text, state)

... where the state is the index of the match to be returned. This is probably
to match the underlying C readline funcion (ahhh, the world before dynamically
allocated lists).

Considering that we're calling the match function N+1 times where N is the
number of autocompletion matches we might as well cache the results. The
autocompleter should have a match function that returns a list of matches
anyway.