Changing HiddenServicePort to be a three value tuple
Our Controller hidden service method's HiddenServicePort was a string. This is
clunky because it left parsing that string to our caller. The string might be
three different formats, anything from a simple port to two ports and an
address.

Our get_hidden_service_conf() now provides a list of three value tuples...

  "HiddenServicePort": [
    (port, target_address, target_port),
  ]

This both normalizes and means no parsing necessary.