Dropping minor tweaks
Just a couple non-impactful tweaks that seem to have slipped in. First is an
unused 'oldstdout' (probably there while debugging) and the second is replacing
io.StringIO with a 'from' import (I liked it a bit better before for uniformity
with the other imports).

I'm also spotting a tweak where Foxboron replaces OrderedDict in the example
test with a normal dict (reverting an earlier change of his). There's another
ticket though that concerns ordering.