Simplifying _get_tests()
Our module_prefixes fall into two camps...

  * It's... well, a module prefix. It'll be a prefix of our import name.
  * It's an exact match for a test. Our import name will be a prefix of it.

No need to strip off the last component and try again. We can do this in one
pass. One thing I noticed though is that we don't check to see if tests exist.
We rely on a try/catch for that. Might be able to wrap it into this helper -
we'll see.