Read multiple descriptors from files with bridge descriptors
Usually descriptors have a keyword that indicates the end of the descriptor.
Sanitized bridge descriptors, however, omit the keyword we check for so they
have no defined ending. This caused our parser to treat the whole file as one
big bridge descriptor.

To fix this we're now splitting these up by their @type annotation. This is a
bit of a hack, but in practice metrics is the only place that provides
sanitized bridge descriptors so should be fine.