Read until 'router-digest' instead of '@type'
Karsten makes the great point that while sanitized bridge descriptors might not
have a 'router-signature' line they instead have a 'router-digest'. Better to
use this to determine the dividing point between descriptors rather than the
@type annotation (which honestly was a bit hacky).

During this discovered that the example we use in our unit tests for a bridge
descriptor doesn't seem to be real. I wasn't able to find it in Collector, so
swapping us over to one that does.