Ensure the 'compare_flags' test gives reliable results.
The call to queries.items() causes the issue by returning an unordered set of
key/value pairs. When the call returns the queries in an unexpected order, the
mocked Query.run() function assigns the network status documents in the wrong
order to the authorities. Using an ordered dictionary fixes the problem.

This commit closes ticket 16299.

Other small changes are a PEP8 fix, a grammar fix, and the use of
assertItemsEqual instead of assertEquals. According to the Python
documentation [0], assertItemsEqual is the Python 2.7 equivalent of
assertCountEqual in Python 3.

[0] https://docs.python.org/2.7/library/unittest.html#unittest.TestCase.assertItemsEqual