Test that tor's config options match the man page
Adding an integ test to check that the configuration options tor provides
matches the man page. This way when new additions are made to tor we'll have a
nice reminder to document them in the manual as well.

In doing this ran into a number of minor issues with the man page. Shot Nick
some patches...

  https://trac.torproject.org/projects/tor/ticket/17666
  https://trac.torproject.org/projects/tor/ticket/17665