Adjusting benchmark for parsing with compression
Earlier I adjusted Stem's benchmarks for when calling parse_file() rather than
the DescriptorReader. This was a ~5% improvement but I didn't have a number for
compression so I left it as-is. This, however, made it look like the
compression delta for python is particularly bad. Subtracting the flat overhead
for the DescriptorReader from that benchmark...

  uncompressed consensus: 876.09 -> 865.72 (-10.37ms)
  compressed consensus: 913.12 -> 902.75