Key fallback directories on fingerprint
Internal detail, but in our config file we keyed fallback directories on their
nickname. This made the config more readable but is a no-go since nicknames are
non-unique. At present tor has two Unnamed fallback directories and as a result
we could only have information about one.

This is all internal details. Only impact for users is that we're no longer
missing one of the fallbacks.