Move event_summary() to control module
Gave this a bit more thought and I was wrong to suggest putting this in
response/events.py. The response module is internal - we don't tell users it
exists so we shouldn't put callable functions in there. The control module
already includes an enum of all the event types so this actually isn't such a
bad fit.