Unit tests fail when no address is available
When expanding the 'private' alias for exit policies the last entry is for our
pubic address. When python is unable to determine one get_config_policy()
excludes this last entry, in turn causing our unit tests to fail...

  https://trac.torproject.org/projects/tor/ticket/20476

Being more careful to treat this part of the policy as being optional.