Speed up controller's test_reattaching_listeners integ test
Test was decent, but overly extensive for the simple thing it aimed to assert.
This test is supposed to simply check that when we reattach to tor we resume
getting events we were previously subscribed for.

Narrowing the test to be just about that and use events we can trigger rather
than waiting for BW events to be emitted (which is the main cause of this
test's long runtime). On my netbook this drops the test from 5.8s to 104 ms
(98% faster).