Don't clone immutables when parsing descriptors
Avoiding an unnecessary copy call on types which are immutable. Seems even for
empty lists and dictionaries calling the constructor rather than copy() is a
tad faster.

On my wee little netbook this speeds up reading microdescriptors by ~5% when
there's validation and speeds our integ tests a bit...

  Before
  descriptor.microdescriptor...                        success (15.38s)
  descriptor.networkstatus...                          success (23.25s)

  After
  descriptor.microdescriptor...                        success (13.96s)
  descriptor.networkstatus...                          success (22.77s)

This should help all descriptor reading though when validation is disabled this
is boon is only when accessing attributes.