Avoid parsing malformed lines multiple times
Descriptor lines usually set multiple attributes. When we've read a line for
one of those attributes there's no point in reading it again for another.

This has no noticeable impact I can tell on performance (multiple calls to a
variety of related attributes in a malformed descriptor is very much an edge
case). Just doing this because when it happens it's clearly wasted work.