Add an is_expired() method to certificates
Method to check if the certificate is presently expired or not. We should *not*
consider a certificate to be invalid because it's expired. Stem is used to
parse past descriptors just as well as current ones. Good thing to be able to
easily check but doesn't belong in a validate() function.