Drop sign_descriptor_content() testing helper
This helper wasn't used by our tests any longer, and on reflection I can't make
the router-signature it generates pass validation. There's one clear bug that
slipped in a while ago: "b'b'.join()" should be "b'\n'.join()". But correcting
that didn't make the function work.

For now just dropping the broken helper. It would be sweeeet to be able to sign
the descriptors we create but that's gonna take a bit of work.