Unit test AUTHDIR_NEWDESC events
Thanks to teor we examples of AUTHDIR_NEWDESC events, so can unit test our
parsing of them. Possibly just in time for them to be removed. :P

  https://trac.torproject.org/projects/tor/ticket/22377

That's fine. Until the event's gone from all non-deprecated tor versions we'll
continue to support them so still good to have coverage here.