Testing failed for higher python3 versions
When testing with python 3.6 our tests fail to run at all with...

  module 'unittest' has no attribute '__version__'

With python 3.2 __import__ worked as we expected but in 3.6 running
"__import__('unittest.mock')" provides the top unittest module rather than
mock.

Turns out the __import__ help info asks callers to never use it, and use
importlib instead. Swapping over fixes this, and is still compatible with
python 2.7.