Parallelize test's static analysis checks
This alone made attending PyCon well worth it. For months I've been struggling
to speed up our tests but python's GIL keeps getting in the way. Turns out the
builtins include a module *specifically* for this...

  https://docs.python.org/3/library/multiprocessing.html#module-multiprocessing

Running our static checks (pyflakes and pycodestyle) in the background. On my
desktop this drops the runtime of our unit tests from 10.8s to 8.9s (18%
faster). I was expecting to get more but this is a pretty old dual-core system
so probably would get more elsewhere.