Exclude hash constant from signatures
Holy crap that was a pita. Oh well, got it. Python's cryptography module
appends a constant indicating the hashing algorithm it uses whereas tor
doesn't. This caused validation of descriptor signatures we add to fail
with...

  ======================================================================
  ERROR: test_descriptor_signing
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "/home/atagar/Desktop/stem/test/require.py", line 58, in wrapped
      return func(self, *args, **kwargs)
    File "/home/atagar/Desktop/stem/test/unit/descriptor/server_descriptor.py", line 260, in test_descriptor_signing
      RelayDescriptor.create(sign = True)
    File "/home/atagar/Desktop/stem/stem/descriptor/server_descriptor.py", line 860, in create
      return cls(cls.content(attr, exclude, sign, private_signing_key), validate = validate, skip_crypto_validation = not sign)
    File "/home/atagar/Desktop/stem/stem/descriptor/server_descriptor.py", line 808, in __init__
      raise ValueError('Decrypted digest does not match local digest (calculated: %s, local: %s)' % (signed_digest, self.digest()))
  ValueError: Decrypted digest does not match local digest (calculated: 3021300906052B0E03021A05000414FDE1A32B71718BFEBD610E76B6340EB4795A554A, local: FDE1A32B71718BFEBD610E76B6340EB4795A554A)

  ----------------------------------------------------------------------

Note that they *do* actually match. Only trouble is that the former is prefixed
by the following constant...

  3021300906052B0E03021A05000414

Took a full day of head scratching but finally got a hint about what this is
thanks to...

  https://crypto.stackexchange.com/questions/8780/verify-signature-failed

Cryptography embeds a constant indicating the hash algorithm used (in this case
sha1). Unfortunately this is baked pretty deep so the only way to exclude this
is mocking out part of its internals. I'll shoot upstream a request in a bit to
allow this to be toggled.