Authentication failures could cause incorrect response
Oops. To best ensure we performed post-authentication activities I hacked it
directly into the BaseController, but these didn't take into account that
authentication could fail. The post-auth requests then failed because tor
closes the socket.

Caught thanks to daftaupe on...

  https://trac.torproject.org/projects/tor/ticket/22679

While troubleshooting this I ran into another bug that might be an issue with
tor. When authentication failed we called 'QUIT'. Tor will sometimes respond
with 'closing connection' and other times just hang. Reached out to Nick to see
if this is something we'd care to address, but on Stem's side now simply
omitting a QUIT call if unauthenticated.