Ensure port used by process tests is unused
When running in a loop catalyst occasionally runs into...

  OSError: Process terminated: Failed to bind one of the listener ports.
  https://trac.torproject.org/projects/tor/ticket/22902

First guess is that maybe multiple tests pick the same random port, or
something that's coincidently being used by the system. Tried to use socket's
connect_ex() for this...

  https://stackoverflow.com/questions/19196105/python-how-to-check-if-a-network-port-is-open-on-linux

But doing so causes exactly the binding failures we're attempting to avoid
(even with a sleep after closing). Weird. Just checking for process instead.