Move get_torrc_entries() into runner
This helper is only used to translate targets into their extra torrc options.
Rest of the torrc stuff resides in the runner so might as well put this there
too.

Not really settled on this yet but presently trying to make our run_tests.py
easier to follow. I'll circle back to do the same for runner.py afterward.