Test mistakenly connecting to SocksPort
While setting up a relay I accidently had nyx connect to the ORPort rather than
the SocksPort. This caused nyx to hang.

Clearly that could be better. Connecting to the wrong port will be a pritty
common mistake. I hoped to repro this with an integ test connecting to the
SocksPort but no luck - this doesn't hang. Oh well. More investigation needed
to get a repro, but at least we got a nice test out of it.