Add DaemonTask class
To improve our test runtime we started using the multithreading module to
parallelize. It's clunky, but with this little helper it's easy to split
off long-running tasks.

This can improve more than just our tests. First candidate that comes to
mind is nyx's resource usage fetches so it doesn't block the draw loop.
Either way, having a tested library class is better than keeping this
hidden in our test helpers.