Allow cached manual to be recreated when deleted
Ensuring Manual.from_cache() raises an IOError rather than a sqlite error when
loading a file that doesn't exist. This should never arise for users when
loading our local copy, but it's handy for cache_manual.py to handle that so we
can delete and recreate the cache more easily.