Provide a connection with stem.manual.database
In dabbling with this more I've been finding the context manager enforced by
stem.manual.database() annoying. When on an interactive prompt I want a method
to quickly get a no-hassle database connection - not a cursor that's only
usable within a 'with' block.

Reason I provided a cursor was to ensure callers had read-only access.
However, just checked and when the database file is read-only (as our
cached_tor_manual.sqlite should be when installed) attempting to write
to it behaves how I'd expect...

  >>> import stem.manual
  >>> conn = stem.manual.database()
  >>> conn.execute('CREATE TABLE new_table(name TEXT PRIMARY KEY, description TEXT)')
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  sqlite3.OperationalError: attempt to write a readonly database