Replace database function with query
On reflection what callers will find useful is an easy function to make sqlite
queries - not a function to get a database connection. Initially I was wary of
this due to being unable to know when to close the connection, but from the
looks of it when accessed in a read-only fashion this doesn't really matter.