Run test_launch_tor_with_timeout in serial
Huh. Works on my laptop but on my desktop one of the tests are failing. I
suspect the issue is that running it at the same time we're running others
in parallel starves the test for resources, causing it to take too long.

Test reliability is far more important than a tiny bit of runtime. On its
own this test is pretty quick anyway. :)

  ======================================================================
  FAIL: test_launch_tor_with_timeout
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "/home/atagar/Desktop/stem/stem/util/test_tools.py", line 150, in <lambda>
      self.method = lambda test: self.result(test)  # method that can be mixed into TestCases
    File "/home/atagar/Desktop/stem/stem/util/test_tools.py", line 212, in result
      test.fail(self._result.msg)
  AssertionError: Test should have taken 0.05-3 seconds, took 3.4 instead

  ----------------------------------------------------------------------