Avoid trace logging when unused
If a tree falls in an empty forest does it make a sound? Constructing trace
runlevel log messages in stem.socket is absolutley *killing* performance on
busy relays. As in pegging cpu to the point of killing the tor process.

... ouch.

Only constructing trace log messages if someone's actually listening for them.