Remove bytes normalization from recv_message
Sockets always provide bytes, but we still normalized the type because tests
and a few methods like ControlMessage.from_str() provided it a StringIO
instead. This is silly. We can provide BytesIOs instead. Our recv_message is in
the critical path so it's important to keep it as minimal as possible.

This might break our tests for python 2.6. Not sure - don't have that python
verison and getting it on modern systems is a pita. I'll check later. 2.6 has
the io module, but unclear if it has BytesIO. If so we might need to tweak our
imports later.