Close control socket faster when under load
When receiving a high volume of traffic (such as DEBUG events on a busy relay)
our controller can take seconds or even minutes to close. This is because we
send a QUIT signal, then wait for the response which is stuck behind
potentially thousands of events.

From what I can tell there's little point in issuing a QUIT so dropping that,
and setting a time limit for processing backlogged events.