Cache 'GETINFO fingerprint' when not a relay
Probably the most requested GETINFO parameter is our fingerprint. We cache
fingerprint values, but not exceptions causing any request when not a relay to
go to tor.

This is unnecessary. We can detect if we become a relay so caching the 'not a
relay' response to avoid having callers (like Nyx) accidently hammer tor.