Skip running async tests under python 2.6
Turns out importlib was added in python 2.7...

  Traceback (most recent call last):
    File "./run_tests.py", line 9, in <module>
      import importlib
  ImportError: No module named importlib

This means we're losing quite a bit of test coverage under python 2.6 but...
screw it. Guido rolled his eyes when I mentioned we even offered 2.6 support.
It's been dead a long time.

We'll continue to support python 2.6 until stem 2.0, but not gonna bend over
backwards for it. This change will let us invoke our tests again.