Read directly from socket with RelaySocket
Ick, struggled with this for a while. The file's read() retrieves data until a
given amount of data has been reached or the socket closed, whereas recv simply
reads whatever is available. Struggled for a while to get read() to behave like
recv() but no luck. Just having the RelaySocket read directly from the socket
instead...

https://stackoverflow.com/questions/6946629/can-i-get-a-socket-makefile-to-have-the-same-read-semantics-as-a-regular-file