Use minimum valid circuit id by default for CREATE_FAST
To work a CREATE_FAST cell must specify a circuit id. Endosome picks one based
on the link version with a similar comment to the one we're including here.
Endosome also has an 'is_initiator' flag which I'm omitting right now since I
suspect that's only relevant if we attempt to implement *being* a relay which
isn't in the cards right now.