Replace pack function with a method
One sucky thing about python: no method overriding. We can have an instance
method or class method called pack, but not both.

Trivial to get around by renaming one of them but on relfection I suspect an
instance method is where we'll want to go in the long run anyway. Lets run with
this for a bit to see how it goes.