Stacktrace when tests call is_accessible after stopping
Huh. I can't repro this failure and I'm unsure how it arose, but our jenkins
tests failed a test run with...

  ======================================================================
  ERROR: test_attachstream
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/require.py", line 57, in wrapped
      if condition():
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/require.py", line 68, in _can_access_controller
      return test.runner.get_runner().is_accessible()
    File "/srv/jenkins-workspace/workspace/stem-tor-ci/test/runner.py", line 341, in is_accessible
      return Torrc.PORT in self._custom_opts or Torrc.SOCKET in self._custom_opts
  TypeError: argument of type 'NoneType' is not iterable

This should only arise if we haven't called start() on our Runner, or called
its stop() method before the test ran. In either case there's no doubt a second
bug here but for the moment just making this obvious fix. All callers expect
the opts to be a collection, so no reason not to initialize it to one.