Provide all cells circuits reply with
Oops, callers of send() on our circuit only got the first cell sent in
repsonse. These are fixed sized cells, so when the response is larger
than 509 bytes (such as in the case of descriptors) multiple will be
received.

It's important that this returns a list rather than yielding a generator
because we need to ensure we issue recv() calls to drain messages off the
socket. Otherwise the next call will hang.