Move Directory into its own module
Directory authorities started as a small, simple constant so it made sense to
co-locate them in the remote module (the sole spot they were used). But they've
grown. With the addition of fallback directories and from_remote() parsers
they're now the bulk of the remote module and way past the point where they
deserve their own module.

Adding a stem.directory module with aliases for backward compatability. Also
shortening the names. With this the module is a *lot* nicer to use. For
instance to list the authorities we're going from...

  for authority in stem.descriptor.remote.DirectoryAuthority.from_cache():
    print(authority.nickname)

... to...

  for authority in stem.directory.Authority.from_cache():
    print(authority.nickname)

Quite a few other improvements I'd like to follow this up with, but starting
with just a simple move.