ConfChangedEvent could not represent multiple config values
Great catch from dmr. The ConfChangedEvent's config was a simple {str => str}
mapping, but some tor configuration options can have multiple values. For
instance...

  650-CONF_CHANGED
  650-ExitPolicy=accept 34.3.4.5
  650-ExitPolicy=accept 3.4.53.3
  650 OK

We can't change the event's config attribute without breaking backward
compatability, but that's ok. Fiddling with this event a bit it's usage
was actually kinda clunky since we bundled config unsetting with
modifications. As such, splitting the 'config' attribute into a 'changed' and
'unset'. This time with changed as a {str => list} mapping.

Here's a little demo of what the new usage might look like...

  def print_config_changes(event):
    for key, values in my_event.changed:
      print('We changed %s to %s' % (key, ', '.join(values))

    for key in my_event.unset:
      print('We removed our configuration of %s' % key)

  my_controller.add_event_listener(print_config_changes, EventType.CONF_CHANGED)