Incorrect circuit digest for higher protocol versions
I only got stem.client working for link protocol 3 when cobbling it together.
We attempt to conform with the spec for higher protocol versions but it's
mostly untested.

While looking into #26060 I realized that we're using the wrong length for cell
headers when the link protocol is higher than three. Link v4 and greater use a
long for circuit identifiers rather than a short...

  circ_id (4 bytes) + command (1 byte) = 5 bytes